requirements.txt
Flask==2.3.2
slack-bolt==1.18.0
openai==0.27.8
python-dotenv==1.0.0